finalized t profiler, prelim test looks ok, needs test full run

+++ radial-spec/radial-spec-driver.sh
@@ -60,7 +60,7 @@ export BG_REGION=${specdir}/${BG_REGION_ID}.phy.reg
 export PS_REGION=${specdir}/${PS_REGION_ID}.phy.reg
 export bgid=${BG_REGION_ID}            # FIXME: refactor redundancy
 
-export LOG_MASTER_FILE="${dir}/${specdir}/conf/${CLNAME}-run-${fitid}-iter-master.tab"
+export LOG_MASTER_FILE="${dir}/${specdir}/conf/${CLNAME}-run-${fitid}-radial-master.tab"
 
 # will need image coordinate point source directory
 cp ${startdir}/${ANALYSIS_DIR}/analysis/${PS_REGION_ID}.im.reg ${specdir}/
@@ -142,36 +142,29 @@ fi
 # initialize iterations
 
 export iter=1
-export r_old=10000.0
-export reached_r_tolerance=0
+export r_old=0.0
 
-export r=$r_init
+export r=$r_min
 export r_phy=$(echo "scale=6;$r*20.0" | bc)
 
 export rcore=$(echo "scale=6;$core_frac*$r" | bc)
 export rcore_phy=$(echo "scale=6;$rcore*20.0" | bc)
+reached_r_max=`echo "if($r > $r_max) 1" | bc`
 
-echo $rcore $rcore_phy
-
-init_log_master $LOG_MASTER_FILE
+echo "initial radius: " $r
+echo "core radius: " $rcore $rcore_phy
+echo "max radius: " $r_max
+echo "r step: " $r_step
+echo "reached r_max: " $reached_r_max
+init_tprofile_log_master $LOG_MASTER_FILE
 
 ######################################################################
 # iterator
 
-while [[ $iter -le $max_iter && $reached_r_tolerance -ne 1 ]]; do
+while [[ $reached_r_max -ne 1 ]]; do
 
     echo "######################################################################"
-    echo "iteration :: " $iter "current val :: " $r " old val :: " $r_old "diff :: " $r_diff, "tolerance reached :: " $reached_r_tolerance
-
-    is_positive=`echo "if($r > 0.0) 1" | bc `
-
-    if [[ $is_positive -eq 0 ]]
-    then
-        echo -e "\n** error: current value for aperture radius is <= 0.0, r=$r"
-        echo -e "*** error in script: $0\n"
-        cd $startdir
-        exit 1
-    fi
+    echo "iteration :: " $iter "current val :: " $r "step :: " $r "maximal :: " $r_max
 
     ######################################################################
     # write region files
@@ -179,7 +172,7 @@ while [[ $iter -le $max_iter && $reached_r_tolerance -ne 1 ]]; do
     shape="circle"
 
     rpad=$(echo $r | bc -l | xargs printf "%1.0f") # round
-    export rpad=`printf "%03d" $rpad`                     # zero pad, need export
+    export rpad=`printf "%03d" $rpad`              # zero pad, need export
     spectrumid="run-$fitid-iter-r-$rpad"           # identifier for spectral products
 
     coordsystem="wcs"
@@ -259,16 +252,15 @@ while [[ $iter -le $max_iter && $reached_r_tolerance -ne 1 ]]; do
             then
                 echo -e "${specdir}/${instrument}-${spectrumid}.rmf does not exists - linking it to preexisting"
 
-                if [[ ! -e ${specdir}/${instrument}.rmf ]]
+                rmfile=${bgspecdir}/${instrument}-cluster-man-01.rmf # use a backup rmf
+                if [[ ! -e $rmfile ]]
                 then
-                    echo -e "\n** error: ${specdir}/${instrument}.rmf does not exists here!"
+                    echo -e "\n** error: $rmfile does not exists here!"
                     echo -e "*** error in $0\n"
                     cd $startdir
                     exit 1
                 fi
-                ln -s ${specdir}/${instrument}.rmf ${specdir}/${instrument}-${spectrumid}.rmf
-
-                # ln -s ${bgspecdir}/${instrument}.rmf ${specdir}/${instrument}-${spectrumid}.rmf
+                ln -s $rmfile ${specdir}/${instrument}-${spectrumid}.rmf
             fi
         done
     fi
@@ -292,15 +284,15 @@ while [[ $iter -le $max_iter && $reached_r_tolerance -ne 1 ]]; do
             then
                 echo -e "${specdir}/${instrument}-${spectrumid}.arf does not exists - linking it to preexisting"
 
-                if [[ ! -e ${specdir}/${instrument}.arf ]]
+                arfile=${bgspecdir}/${instrument}-cluster-man-01.arf # use a backup arf
+                if [[ ! -e $arfile ]]
                 then
-                    echo -e "\n** error: ${specdir}/${instrument}.arf does not exists here!"
+                    echo -e "\n** error: $arfile does not exists here!"
                     echo -e "*** error in $0\n"
                     cd $startdir
                     exit 1
                 fi
-                ln -s ${specdir}/${instrument}.arf ${specdir}/${instrument}-${spectrumid}.arf
-                # ln -s ${bgspecdir}/${instrument}.arf ${specdir}/${instrument}-${spectrumid}.arf
+                ln -s $arfile ${specdir}/${instrument}-${spectrumid}.arf
             fi
         done
     fi
@@ -308,16 +300,19 @@ while [[ $iter -le $max_iter && $reached_r_tolerance -ne 1 ]]; do
     ######################################################################
     # do the oot subtraction
 
-    echo "Running oot subtraction: "
-    export here=`pwd`
-    cd $specdir
-    echo "doing oot subtraction"
-    subtract_oot_spec ${instrument}-${spectrumid}.pha ${instrument}-${spectrumid}-oot.pha
+    if [[ $EXTRACT_SRC_SPEC -eq 1 ]]
+    then
+        echo "Running oot subtraction: "
+        export here=`pwd`
+        cd $specdir
+        echo "doing oot subtraction"
+        subtract_oot_spec ${instrument}-${spectrumid}.pha ${instrument}-${spectrumid}-oot.pha
 
     # comment for debugging
-    rm ${instrument}-${spectrumid}-oot.pha ${instrument}-${spectrumid}-oot.im ${instrument}-${spectrumid}-oot.pha.orig
+        rm ${instrument}-${spectrumid}-oot.pha ${instrument}-${spectrumid}-oot.im ${instrument}-${spectrumid}-oot.pha.orig
 
-    cd $here
+        cd $here
+    fi
 
     ######################################################################
     # do the spectral fit
@@ -385,31 +380,26 @@ while [[ $iter -le $max_iter && $reached_r_tolerance -ne 1 ]]; do
     fi
 
     ######################################################################
-    # update the radius
+    # write into the master
 
-    r_old=$r
+    if [[ $DO_SPECTROSCOPY -eq 1 ]]
+    then
+        aper_results=`read_aper_result_file ${spectrumid}/${CLNAME}-${spectrumid}.aper`
 
-    # r=$(echo "scale=6;$r*1.2" | bc) # dummy for debug
-    ${codedir}/py/t_to_r.py ${spectrumid}/${CLNAME}-${spectrumid}.result | tee ${spectrumid}/${CLNAME}-${spectrumid}.aper
-
-    r=`egrep "\brfit_ang\b" ${spectrumid}/${CLNAME}-${spectrumid}.aper | awk '{print $2}'`
-    r_phy=$(echo "scale=6;$r*20.0" | bc)
-
-    rcore=$(echo "scale=6;$core_frac*$r" | bc)
-    rcore_phy=$(echo "scale=6;$rcore*20.0" | bc)
-
-    r_diff=$(echo "scale=6;$r-$r_old" | bc)
-    r_diff=`echo ${r_diff#-}`   # get the absolute value
-    reached_r_tolerance=`echo "if($r_diff <= $r_tolerance) 1" | bc`
+        echo $fitid $iter $r $aper_results >> $LOG_MASTER_FILE
+        mv formated_pars.out ${CLNAME}-${spectrumid}.fpars
+    else
+        echo $fitid $iter $r >> $LOG_MASTER_FILE
+    fi
 
     ######################################################################
-    # write into the master: note: the current line corresponds to the
-    # spectrum fitted in the (now) r_old aperture
+    # prepare new aperture
+    iter=$((iter + 1))
+    r_old=$r
+    r=$(echo "scale=6;$r+$r_step" | bc)
+    echo $r_old $r_step $r "@@@@@@@@@@@@@@@@@@@@"
 
-    aper_results=`read_aper_result_file ${spectrumid}/${CLNAME}-${spectrumid}.aper`
-
-    echo $fitid $iter $r_old $r $r_diff $aper_results >> $LOG_MASTER_FILE
-    mv formated_pars.out ${CLNAME}-${spectrumid}.fpars
+    reached_r_max=`echo "if($r > $r_max) 1" | bc`
 
     ######################################################################
     # finished iteration step
@@ -419,32 +409,31 @@ while [[ $iter -le $max_iter && $reached_r_tolerance -ne 1 ]]; do
     cd $dir                     # returns to analysis subdir
     pwd
 
-    iter=$((iter + 1))
-
     echo "Done:"
     echo "iteration :: " $iter " for r val :: " $r_old " r for the next iteration :: " $r >>  ${specdir}/${spectrumid}/${CLNAME}-${spectrumid}.aper
     echo "######################################################################" >> ${specdir}/${spectrumid}/${CLNAME}-${spectrumid}.aper
     echo
 
-done
-
-mv ${specdir}/${spectrumid} ${specdir}/${spectrumid}-final
+done                            # iteration end
+######################################################################
 
 ######################################################################
-# do area correction calculation
+# do area correction calculation - NOT needed for radial temperature
+# profiles
 
-if [[ $CALCULATE_AREACORR -eq 1 ]]
-then
-    cd ${specdir}/${spectrumid}-final
-    if [[ $EXCLUDE_CORE -eq 0 ]]
-    then
-        area-correction-factors.sh "$r_old"
-    else
-        rcore_old=$(echo "scale=6;$core_frac*$r_old" | bc)
-        area-correction-factors.sh "$rcore_old" "$r_old"
-    fi
+# if [[ $CALCULATE_AREACORR -eq 1 ]]
+# then
+#     cd ${specdir}/${spectrumid}-final
+#     if [[ $EXCLUDE_CORE -eq 0 ]]
+#     then
+#         area-correction-factors.sh "$r_old"
+#     else
+#         rcore_old=$(echo "scale=6;$core_frac*$r_old" | bc)
+#         area-correction-factors.sh "$rcore_old" "$r_old"
+#     fi
+# fi
 
-fi
+
 
 ######################################################################
 # exit
